+++ src/Components/LandingContainer/LandingContainer.jsx
@@ -1,8 +1,8 @@
-import { useReducerContext } from "../../context/useReducer.context";
+import { useFilterReducerContext } from "../../context/useFilterReducer.context";
 import "./LandingContainer.css";
 
 export const LandingContainer = ({ children }) => {
-  const { sidebar } = useReducerContext();
+  const { sidebar } = useFilterReducerContext();
   return (
     <section className={`landing-container ${sidebar ? "position-fixed" : ""}`}>
       {children}
