+++ src/Pages/LandingPage/LandingComponents/FeaturedContainer/CategorySection/CategorySection.jsx
@@ -1,10 +1,10 @@
 import axios from "axios";
 import { useState, useEffect } from "react";
 import { Loader } from "../../../../../Components";
-import { useReducerContext } from "../../../../../context/useReducer.context";
+import { useFilterReducerContext } from "../../../../../context/useFilterReducer.context";
 
 export const CategorySection = () => {
-  const { loading, dispatch } = useReducerContext();
+  const { loading, dispatch } = useFilterReducerContext();
 
   const [categoryData, setCategoryData] = useState([]);
