+++ src/Pages/ProductPage/ProductComponents/CardContainer/CardContainer.css
@@ -3,7 +3,7 @@
 }
 
 .card-container.product-card {
-  width: 35rem;
+  width: 30rem;
   justify-content: space-between;
 }
 
@@ -27,8 +27,8 @@
 }
 
 .product-img {
-  width: 35rem;
-  height: 35rem;
+  width: 30rem;
+  height: 30rem;
 }
 
 .card-container .card-title {
@@ -71,29 +71,12 @@
 
 @media screen and (max-width: 1300px) {
   .card-container.product-card {
-    width: 30rem;
     margin: 5px;
     min-height: 55rem;
-  }
-
-  .product-img {
-    width: 30rem;
-    height: 30rem;
+    margin-top: 2rem;
   }
 
   .out-of-stock {
       font-size: 4rem;
-  }
-}
-
-@media screen and (max-width: 1300px) {
-  .card-container.product-card {
-    width: 35rem;
-    margin-top: 2rem;
-  }
-
-  .product-img {
-    width: 35rem;
-    height: 35rem;
   }
 }
